requirements.txt
torch==1.9.0+cu102
torchvision==0.10.0+cu102
Pillow==8.3.1
requests==2.25.1
pandas==1.3.0
tensorboard==2.5.0
streamlit==0.84.1